refactor : CommonSecurityTest -> CommonAcceptanceTest로 변경
